Video RTP source node now uses explicit subprocesses
The ffmpeg-python bindings are now removed from the video_rtp source node,
in favour of explicit ffmpeg subprocess. This clears a third-party dependency.
Additionally, the extra ffmpeg decoding process to be used in conjunction with
the local RTP client is removed, as the current ffmpeg subprocess directly
writes the received frames on the standard output to be fetched by the node
source function.