Node typing annotations
This is a tentative structure that allows to annotate node in and out types. A
number of default payloads are defined, that can be used when creating a
message. Nodes can now be created by specifying the type of message they expect
to receive as input, and the type they produce as output.