This commit include 2 major changes:
1. Node init signature changed (breaking)
2. Logging facility

Node signature changed
======================
When creating a custom node, it is not necessary to hardcode the node type and
use it in the BaseNode constructor. This information will be automatically
provided to the node constructor within the kwargs parameter, alongside pipe
and node names. This has 2 advantages: it removes the need to write the node
type both in the configuration and in the code, and it offers the node its own
name at creation time. With this change, kwargs have to be included in the
signature of every node, and they should be passed to the superclass
constructor.

Logging facility
================
Juturna now offers basic logging functionalities, provided as:

- a root, library-level logger called jt;
- individual loggers for any active pipelines and their nodes within, all of
  which are child loggers of the root logger.

Child loggers are created with the names of their corresponding entity, and
their levels can be selectively set. Disabling the root logger will disable
all logging from the library, while disabling the logger of a specific
pipeline will disable the logging of both the pipe and all its nodes.

---------------

A new node was added to the plugin directory, called TranslatorNllb.