Multi-input node structure
The need for nodes capable of accepting data from multiple input nodes requires
the redesign of the core functionalities of the node internals. Nodes now use
synchronised queues to read input data and write output data. After a message
is consumed at the input, a manager is responsible for syncing it with any
other messages currently stored, and eventually producing a Batch of messages
to the output queue.